extra test futures & refactor
